Really fix security vulnerability
List action didn't sanitize has_many association
name properly, which was directly used in a call
to #send. A malicious user could specify delete
or destroy as a has_many association name which
would cause a record to be destroyed. This patch
checks that params[:children] is a valid name.

Signed-off-by: Alexey Borzenkov <snaury@gmail.com>